+++ src/Thread/TaskQueue.h
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLToolKit project authors. All Rights Reserved.
  *
- * This file is part of ZLToolKit(https://github.com/xiongziliang/ZLToolKit).
+ * This file is part of ZLToolKit(https://github.com/xia-chu/ZLToolKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -59,7 +59,7 @@ public:
         _queue.pop_front();
         return true;
     }
-    uint64_t size() const{
+    size_t size() const{
         lock_guard<decltype(_mutex)> lock(_mutex);
         return _queue.size();
     }
