+++ src/Util/SSLUtil.cpp
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLToolKit project authors. All Rights Reserved.
  *
- * This file is part of ZLToolKit(https://github.com/xiongziliang/ZLToolKit).
+ * This file is part of ZLToolKit(https://github.com/xia-chu/ZLToolKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -129,7 +129,7 @@ shared_ptr<EVP_PKEY> SSLUtil::loadPrivateKey(const string &file_path_or_data, co
 
     pem_password_cb *cb = [](char *buf, int size, int rwflag, void *userdata) -> int{
         const string *passwd = (const string *)userdata;
-        size = size < passwd->size() ? size :  passwd->size();
+        size = (int)(size < passwd->size() ? size :  passwd->size());
         memcpy(buf,passwd->data(),size);
         return size;
     };
