+++ src/win32/getopt.c
@@ -177,7 +177,7 @@ extern int  strcmp (const char *s1, const char *s2);
 #if defined(__MINGW32_MAJOR_VERSION)
 extern int  strncmp(const char *_Str1,const char *_Str2,size_t _MaxCount);
 #else
-extern int  strncmp(const char *s1, const char *s2, int n);
+extern int  strncmp(const char *s1, const char *s2, size_t n);
 #endif //
 static int my_strlen(const char *s);
 static char *my_index (const char *str, int chr);
